Add files via upload
Thinking about the future and future business issues that might occur, we chose the Spaceship Titanic problem. We know what happened to the Titanic ship and how people lost their lives and the shipping company had to compensate for it, along with their reputation being dragged down to the ground. Similarly, a month ago, the Spaceship Titanic, an intergalactic passenger liner, was launched. The ship left on its inaugural mission carrying emigrants from our solar system to three newly habitable exoplanets circling neighboring stars with approximately 13,000 passengers on board.
The unsuspecting Spaceship Titanic collided with a spacetime anomaly concealed beneath a dust cloud as it rounded Alpha Centauri on its way to its first destination—the scorching 55 Cancri E. Sadly, it experienced the same demise as its namesake. The ship was unharmed, but approximately half of the people were taken to a different dimension. 
In order to prevent the same debacle from happening to the spaceship company we are using the data collected from the spaceship's broken computer system, we are identifying which passengers were carried by the anomaly to aid rescue teams and locate the missing passengers. This would not only help rescue people but also help the company by not having huge lawsuits filed against them and their reputation stay intact. However, to predict this we are first going to understand the data, do an Exploratory Data Analysis then clean the data to eliminate unnecessary variables. Then work on different prediction models to find the optimal solution. 
Anomaly detection is growing in significance as a technique of data analysis and alerts. A value that deviates from the norm by a large enough margin to be considered an anomaly. The development of patterns is a necessary initial step in the detection process, followed by the identification of the units that violate those patterns. We can find this anomaly detection in fraud detection, intrusion detection, and many more. Talking fraud detection, credit card, bank account, and insurance fraud are all prevented with the use of graph-based anomaly detection (GBAD). With the use of behavioral biometrics, which can also detect irregularities in customer spending in real-time, ML systems also facilitate online banking fraud.